requirements/base.txt
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-filter==24.3
dj-database-url==2.2.0
Pillow==10.4.0 